Select menu implemented.

+++ src/wifi_manager.c
@@ -32,11 +32,11 @@ static wifi_manager_status_t s_wifi_status = WIFI_MANAGER_DISCONNECTED;
 #define WIFI_CONNECTED_BIT BIT0
 #define WIFI_FAIL_BIT BIT1
 
-static void event_handler(void* arg, esp_event_base_t event_base, int32_t event_id,
-                          void* event_data) {
+static void
+event_handler(void* arg, esp_event_base_t event_base, int32_t event_id, void* event_data) {
     if (event_base == WIFI_EVENT) {
         if (event_id == WIFI_EVENT_STA_START) {
-            wifi_config_t config = {0};
+            wifi_config_t config = { 0 };
             esp_wifi_get_config(WIFI_IF_STA, &config);
 
             if (config.ap.ssid[0] != '\0' && Config.wifi_on) {
@@ -84,11 +84,13 @@ void wifi_manager_init(void) {
     esp_event_handler_instance_t instance_any_id;
     esp_event_handler_instance_t instance_got_ip;
 
-    ESP_ERROR_CHECK(esp_event_handler_instance_register(WIFI_EVENT, ESP_EVENT_ANY_ID,
-                                                        &event_handler, NULL, &instance_any_id));
+    ESP_ERROR_CHECK(esp_event_handler_instance_register(
+        WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler, NULL, &instance_any_id
+    ));
 
-    ESP_ERROR_CHECK(esp_event_handler_instance_register(IP_EVENT, IP_EVENT_STA_GOT_IP,
-                                                        &event_handler, NULL, &instance_got_ip));
+    ESP_ERROR_CHECK(esp_event_handler_instance_register(
+        IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler, NULL, &instance_got_ip
+    ));
 
     // initialize wifi with default config
     ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_STA));
@@ -126,8 +128,9 @@ esp_err_t wifi_manager_connect_to_ap(const char* ssid, const char* key) {
     ESP_LOGI(TAG, "Connecting to WiFi: \"%s\"", wifi_config.sta.ssid);
     ESP_LOGD(TAG, "    Using Password: \"%s\"", wifi_config.sta.password);
 
-    EventBits_t bits = xEventGroupWaitBits(s_wifi_event_group, WIFI_CONNECTED_BIT | WIFI_FAIL_BIT,
-                                           pdFALSE, pdFALSE, portMAX_DELAY);
+    EventBits_t bits = xEventGroupWaitBits(
+        s_wifi_event_group, WIFI_CONNECTED_BIT | WIFI_FAIL_BIT, pdFALSE, pdFALSE, portMAX_DELAY
+    );
 
     if (bits & WIFI_CONNECTED_BIT) {
         ESP_LOGI(TAG, "Connected to AP.");
@@ -156,10 +159,10 @@ esp_err_t wifi_manager_connect_to_ap(const char* ssid, const char* key) {
  * Scan for WiFi networks and populate a buffer. *count should initially contain the maximum number
  * of networks to scan. After this function returns, it will contain the number of networks found.
  */
-esp_err_t wifi_manager_scan(wifi_ap_record_t* ap_info, size_t* count) {
+esp_err_t wifi_manager_scan(wifi_ap_record_t* ap_info, uint16_t* count) {
     esp_err_t err = ESP_OK;
-    size_t len = *count;
-    size_t ap_count = 0;
+    uint16_t len = *count;
+    uint16_t ap_count = 0;
     memset(ap_info, 0, sizeof(wifi_ap_record_t) * len);
 
     // Start a blocking scan:
@@ -189,7 +192,9 @@ void wifi_manager_disconnect(void) {
     esp_wifi_disconnect();
 }
 
-wifi_manager_status_t wifi_manager_get_status(void) { return s_wifi_status; }
+wifi_manager_status_t wifi_manager_get_status(void) {
+    return s_wifi_status;
+}
 
 int8_t wifi_manager_get_rssi(void) {
     wifi_ap_record_t ap;
@@ -198,5 +203,5 @@ int8_t wifi_manager_get_rssi(void) {
 }
 
 const char* wifi_manager_get_local_ip(void) {
-    return s_wifi_ip_addr_str; 
+    return s_wifi_ip_addr_str;
 }
